+++ action_server/src/sema4ai/action_server/env/__init__.py
@@ -1,11 +1,15 @@
 import logging
+import typing
 
 from sema4ai.action_server._protocols import ArgumentsNamespace, ArgumentsNamespaceEnv
+
+if typing.TYPE_CHECKING:
+    from sema4ai.action_server._rcc import Rcc
 
 log = logging.getLogger(__name__)
 
 
-def handle_env_command(env_args: ArgumentsNamespace):
+def handle_env_command(env_args: ArgumentsNamespace, rcc: "Rcc"):
     import typing
 
     env_command = typing.cast(ArgumentsNamespaceEnv, env_args).env_command
@@ -15,14 +19,7 @@ def handle_env_command(env_args: ArgumentsNamespace):
 
     if env_command == "clean-tools-caches":
         log.info("Clearing tools caches, please wait...")
-        from sema4ai.action_server._download_rcc import download_rcc
-        from sema4ai.action_server._rcc import initialize_rcc
-
-        robocorp_home = None
-
-        with initialize_rcc(download_rcc(force=False), robocorp_home) as rcc:
-            rcc.clean_tools_caches()
-
+        rcc.clean_tools_caches()
         return 0
 
     log.critical(f"Env command not recognized: {env_command}.")