+++ templates/advanced/actions.py
@@ -6,9 +6,9 @@ https://github.com/sema4ai/actions/blob/master/README.md
 
 """
 
-import requests
 from sema4ai.actions import action, Secret
 from pydantic import BaseModel, Field
+import requests
 
 
 class RepositoryInfo(BaseModel):
@@ -18,9 +18,7 @@ class RepositoryInfo(BaseModel):
 
 @action(is_consequential=False)
 def get_repository_commits(
-        github_access_token: Secret,
-        repository_info: RepositoryInfo,
-        limit: int = 10
+    github_access_token: Secret, repository_info: RepositoryInfo, limit: int = 10
 ) -> str:
     """
     Returns commit messages from given repository.
@@ -33,20 +31,21 @@ def get_repository_commits(
     Returns:
         A list of commit messages.
     """
-
     if limit < 1 or limit > 100:
         return "Limit must be between 1 and 100"
 
-    response = requests.get(f"https://api.github.com/repos/{repository_info.owner_id}/{repository_info.name}/commits?per_page={limit}",
-                            headers={
-                                'Authorization': 'Bearer ' + github_access_token.value,
-                                'Content-Type': 'application/x-www-form-urlencoded'
-                            })
+    response = requests.get(
+        f"https://api.github.com/repos/{repository_info.owner_id}/{repository_info.name}/commits?per_page={limit}",
+        headers={
+            "Authorization": "Bearer " + github_access_token.value,
+            "Content-Type": "application/x-www-form-urlencoded",
+        },
+    )
 
     items = response.json()
-    commit_messages = [item['commit']['message'] for item in items]
+    commit_messages = [item["commit"]["message"] for item in items]
 
-    output = '\n'.join(commit_messages)
+    output = "\n".join(commit_messages)
 
     # Pretty print for log
     print(output)